Коллекция всех фактов

+++ app/src/main/java/com/alonsonya/dailyfurnace/data/Furnace.kt
@@ -1,7 +1,7 @@
 package com.alonsonya.dailyfurnace.data
 
 data class Furnace(
-    val furnaceId: Int,
+    val furnaceId: Long,
     val furnaceName: String,
     val furnaceInfo: String,
     val furnaceType: String,
